[LT] Upstream LazyView and view ops IR Nodes
Summary: LazyView is the main class for tracking alias caused by view
ops. The corresponding IR classes for view ops are hand-written now, and
we can switch to code-gen them in future. For certain view ops, they
have a reverse IR class to perform inplace update in the backward
direction on a chain of alias ops.

As part of the future work, we will simplify the logic for LazyView once
the functionalization pass in core is ready to use.

ghstack-source-id: 4cc058acf50abb5c9abcb21a85c23c537a482acd
Pull Request resolved: https://github.com/pytorch/pytorch/pull/69277