fx quant: fix bug with fusion on a module used by multiple nodes
Summary:

Before this PR, the FX quantization fusion logic incorrectly fused
modules which had multiple nodes using them.  For example,

```
class M(torch.nn.Module):
    def __init__(self):
        super().__init__()
        self.conv = torch.nn.Conv2d(1, 1, 1)
        self.bn = torch.nn.BatchNorm2d(1)

    def forward(self, x):
        x = self.conv(x)
        x = self.conv(x)
        x = self.bn(x)
        return x
```

the second conv-bn would be fused, which would change the meaning of
the first call to the conv.

This PR changes the pattern matching logic to disable matching
for any nodes which are not the last node in a pattern and which
refer to modules which have multiple module_uses.

Note that this definition of "module_use" is different from the "use" in the
FX framework.  In FX, a use is referring to which nodes use the output
of a node.  "module_use" refers to how many FX nodes use a module
instance.

Test Plan:

```
python test/test_quantization.py TestFuseFx.test_fusion_turns_off_for_multi_use_intermediate_nodes
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: d811cd7b3ded8b48a3ad1d290a1d672bdd84e5c4
Pull Request resolved: https://github.com/pytorch/pytorch/pull/68178